requirements.txt
black==23.3.0
dynaconf==3.2.1
flake8==6.0.0
pan-os-python==1.11.0
xmltodict==0.13.0